Refactor: relocate 'QuestionsList' component
* Helps to consolidate CSS
